+++ src/main/java/uz/pdp/heymasterapp/repository/AttachmentRepository.java
@@ -8,7 +8,7 @@ import uz.pdp.heymasterapp.entity.Attachment;
 import java.util.List;
 
 @Repository
-public interface AttachmentRepository extends JpaRepository<Attachment, Integer> {
+public interface AttachmentRepository extends JpaRepository<Attachment, Long> {
 
 
   // Attachment findByProfilePhoto(boolean profilePhoto);
